sponsor: A new dedicated sponsor page
For now this simply links to our existing funding channels,
while thanking the remarkable folks at Framework for the 13.
It also alludes to our being open to further hardware/hosting
opportunities, and makes it clear that our current 2 funding
channels aren't going to be the only ones available. Our intent
is still to establish a project-wide-funding and reward scheme.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>